dernier

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: afourcad <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 16:54:12 by afourcad          #+#    #+#             */
-/*   Updated: 2016/11/04 16:54:13 by afourcad         ###   ########.fr       */
+/*   Updated: 2016/11/15 18:26:40 by afourcad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
